feat: adjustments made on the pets list screen
