+++ frontend/src/routes/routes.tsx
@@ -1,6 +1,6 @@
 import { createBrowserRouter, Navigate } from 'react-router-dom'
 import { RootLayout } from '../layouts/RootLayout/RootLayout'
-import { PetsListing } from '../pages/Pets/PetsListing'
+import { PetsListing } from '../pages/Pets/PetsListing/PetsListing'
 import { PetDetail } from '../pages/Pets/PetDetail'
 import { CustomersListing } from '../pages/Customers/CustomersListing/CustomersListing'
 import { CustomerDetail } from '../pages/Customers/CustomerDetail/CustomerDetail'
